add ui kit and edit the 2 forms setup
