+++ src/entities/FormLink/FormLink.styles.ts
@@ -1,3 +1,4 @@
+import { Link } from "react-router";
 import styled from "styled-components";
 
 export const StyledFormLink = styled.div`
@@ -5,4 +6,9 @@ export const StyledFormLink = styled.div`
   justify-content: center;
   align-items: center;
   gap: 1rem;
+`;
+
+export const StyledAuthLink = styled(Link)`
+  text-decoration: none;
+  color: inherit;
 `;
