--- src/shared/ui/AuthFormSubmit/AuthFormSubmitBtn.styles.ts
@@ -1,7 +1,0 @@
-import { Link } from "react-router";
-import styled from "styled-components";
-
-export const StyledAuthLink = styled(Link)`
-  text-decoration: none;
-  color: inherit;
-`;
