+++ src/shared/ui/Input/Input.tsx
@@ -1,0 +1,6 @@
+import { StyledTextField } from "./Input.styles";
+
+const Input = ({ ...props }) => {
+  return <StyledTextField {...props} />;
+};
+export default Input;
